+++ apps/auth-cdn/src/app/resource/dto/resources-query.dto.ts
@@ -1,0 +1,9 @@
+import { IsOptional } from 'class-validator';
+
+export class ResourcesQueryDto {
+  @IsOptional()
+  filename: string;
+
+  @IsOptional()
+  folder: string;
+}